requirements.txt
-i https://pypi.org/simpleatomicwrites==1.4.0; sys_platform == 'win32'attrs==19.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'colorama==0.4.3; sys_platform == 'win32'iniconfig==1.0.0more-itertools==8.4.0; python_version >= '3.5'packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'py==1.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'pytest-dotenv==0.5.2pytest==6.0.0python-decouple==3.3python-dotenv==0.14.0singleton-decorator==1.0.0six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'sqlalchemy==1.3.18toml==0.10.1factory-boy==2.12.0